Add stale building mode
This way the building replication ABM doesn't re-try accessing the
last element of the array over and over again. Benefit may be discussable,
but later it might give notifications to a player if a building is stale.

If a node updates, the on_dignode and on_placenode listener callbacks will
recognize this, and set state to building (or built), re-triggering a
new check if needed.